Change how the GitHub DNS verification secret is set
Was going to use a SecretsManager dynamic reference, but apparently that's not supported on this Route53 resource property (even though this isn't documented anywhere)? ssm-secure dynamic references wouldn't work either (they actually ARE documented to not work here), and SSM parameter types wouldn't work because we can't use SecureString Systems Manager parameter types. So instead, we'll just take the TXT value itself as a parameter and add the necessary quotes. Less secure b/c the value is saved in the Cfn template...but at least it's not in the repo.