WebsiteRedirectStack must now be deployed in us-east-1 region
Its ACM certificate (so its CloudFront distribution can support HTTPS), must be in us-east-1. Since these redirect distros probably won't get a lot of traffic, forcing the whole stack to be in us-east-1 (and not a region closer to users) doesn't seem like it will ever be an issue, and spares creating another stack just for the cert.