Tweak composite health check alarms
Previously, I had separate alarms for when a redirect domain health-check-status was alarming  and 1) main site was alarming, or 2) main site was healthy. The two alarms made sense, but it means the latter would always fire after the former as site was becoming healthy and redirect alarms hadn't updated yet. Getting an alarm right as the site is healing isn't ideal.

Instead, the former alarm is still around, but latter alarm is now for anytime ANY redirect domain is alarming. So two alarms will now be fired when redirect domains break due to site being down, and still only latter will alarm when a redirect domain is unhealthy but main website is fine.